refactor: move validators into vda5050_core::validation namespace
