+++ vda5050_core/CMakeLists.txt
@@ -95,18 +95,19 @@ target_include_directories(vda5050_errors
     $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
 )
 
-# Order Utils
-add_library(vda5050_order_utils SHARED src/order_utils/order_graph_validator.cpp)
+# Validation
+file(GLOB_RECURSE validation_srcs "src/validation/*.cpp")
+add_library(vda5050_validation SHARED ${validation_srcs})
 vda5050_configure_target_alias(
-  TARGET vda5050_order_utils
+  TARGET vda5050_validation
   NAMESPACE ${PROJECT_NAME}::
-  EXPORT_NAME order_utils
+  EXPORT_NAME validation
 )
-target_link_libraries(vda5050_order_utils PUBLIC
+target_link_libraries(vda5050_validation PUBLIC
   ${PROJECT_NAME}::types
   ${PROJECT_NAME}::errors
 )
-target_include_directories(vda5050_order_utils
+target_include_directories(vda5050_validation
   PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
     $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
@@ -162,7 +163,7 @@ target_link_libraries(vda5050_client
   PUBLIC
     ${PROJECT_NAME}::execution
     ${PROJECT_NAME}::types
-    ${PROJECT_NAME}::order_utils
+    ${PROJECT_NAME}::validation
     ${PROJECT_NAME}::logger
 )
 target_include_directories(vda5050_client
@@ -204,7 +205,7 @@ target_link_libraries(vda5050_master
     ${PROJECT_NAME}::json_utils
     ${PROJECT_NAME}::transport
     ${PROJECT_NAME}::execution
-    ${PROJECT_NAME}::order_utils
+    ${PROJECT_NAME}::validation
     ${PROJECT_NAME}::errors
 )
 target_include_directories(vda5050_master
@@ -225,7 +226,7 @@ install(
     vda5050_json_utils
     vda5050_logger
     vda5050_errors
-    vda5050_order_utils
+    vda5050_validation
     vda5050_transport
     vda5050_execution
     vda5050_client
@@ -298,10 +299,6 @@ if(BUILD_TESTING)
   ament_add_gmock(test_json_utils ${json_tests})
   target_link_libraries(test_json_utils vda5050_core::json_utils)
 
-  ament_add_gmock(test_order_utils
-    test/order_utils/test_unit_order_graph_validator.cpp)
-  target_link_libraries(test_order_utils vda5050_core::order_utils)
-
   ament_add_gmock(test_logger
     test/logger/test_unit_default_logger.cpp
     test/logger/test_unit_logger.cpp
@@ -344,11 +341,12 @@ if(BUILD_TESTING)
   )
   target_link_libraries(test_master vda5050_core::master)
 
-  ament_add_gmock(test_master_validation
-    test/master/test_unit_content_validator.cpp
-    test/master/test_unit_pre_send_validator.cpp
+  ament_add_gmock(test_validation
+    test/validation/test_unit_content_validator.cpp
+    test/validation/test_unit_pre_send_validator.cpp
+    test/validation/test_unit_order_graph_validator.cpp
   )
-  target_link_libraries(test_master_validation vda5050_core::master)
+  target_link_libraries(test_validation vda5050_core::validation)
 
   ament_add_gmock(test_layout_loader
     test/layout/test_unit_loader.cpp
