+++ vda5050_core/include/vda5050_core/validation/order_graph_validator.hpp
@@ -16,15 +16,15 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__ORDER_UTILS__ORDER_GRAPH_VALIDATOR_HPP_
-#define VDA5050_CORE__ORDER_UTILS__ORDER_GRAPH_VALIDATOR_HPP_
+#ifndef VDA5050_CORE__VALIDATION__ORDER_GRAPH_VALIDATOR_HPP_
+#define VDA5050_CORE__VALIDATION__ORDER_GRAPH_VALIDATOR_HPP_
 
 #include "vda5050_core/errors/validation_result.hpp"
 #include "vda5050_core/types/order.hpp"
 
 namespace vda5050_core {
 
-namespace order_utils {
+namespace validation {
 
 /// \brief Checks that the nodes and edges in a VDA5050 Order form a valid
 /// graph according to the VDA5050 specification sheet.
@@ -45,7 +45,7 @@ errors::ValidationResult is_valid_update(
   const vda5050_core::types::Order& base_order,
   const vda5050_core::types::Order& next_order);
 
-}  // namespace order_utils
+}  // namespace validation
 }  // namespace vda5050_core
 
-#endif  // VDA5050_CORE__ORDER_UTILS__ORDER_GRAPH_VALIDATOR_HPP_
+#endif  // VDA5050_CORE__VALIDATION__ORDER_GRAPH_VALIDATOR_HPP_
