+++ vda5050_core/test/validation/test_unit_order_graph_validator.cpp
@@ -22,7 +22,7 @@
 #include <vector>
 
 #include "vda5050_core/errors/error_codes.hpp"
-#include "vda5050_core/order_utils/order_graph_validator.hpp"
+#include "vda5050_core/validation/order_graph_validator.hpp"
 
 class OrderValidationTest : public testing::Test
 {
@@ -61,7 +61,7 @@ protected:
 TEST_F(OrderValidationTest, EmptyNodeList)
 {
   auto order = create_order("order_0", 0);
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -80,7 +80,7 @@ TEST_F(OrderValidationTest, UnequalNodesEdges)
 
   order.edges = {create_edge("edge_0", 1, "node_0", "node_1", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -102,7 +102,7 @@ TEST_F(OrderValidationTest, FirstSequenceOfUpdate)
     create_edge("edge_0", 3, "node_0", "node_1", true),
     create_edge("edge_1", 5, "node_1", "node_2", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -124,7 +124,7 @@ TEST_F(OrderValidationTest, UnreleasedFirstNode)
     create_edge("edge_0", 1, "node_0", "node_1", false),
     create_edge("edge_1", 3, "node_1", "node_2", false)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -140,7 +140,7 @@ TEST_F(OrderValidationTest, EvenNodeSequences)
 
   order.nodes = {create_node("node_0", 1, true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -162,7 +162,7 @@ TEST_F(OrderValidationTest, NodeBaseHorizonSeparation)
     create_edge("edge_0", 1, "node_0", "node_1", true),
     create_edge("edge_1", 3, "node_1", "node_2", false)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -182,7 +182,7 @@ TEST_F(OrderValidationTest, OddEdgeSequences)
 
   order.edges = {create_edge("edge_0", 2, "node_0", "node_1", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -201,7 +201,7 @@ TEST_F(OrderValidationTest, DisconnectedGraph)
 
   order.edges = {create_edge("edge_0", 1, "node_0", "node_1", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -219,7 +219,7 @@ TEST_F(OrderValidationTest, DisconnectedEdge)
 
   order.edges = {create_edge("edge_0", 1, "node_0", "node_2", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -241,7 +241,7 @@ TEST_F(OrderValidationTest, EdgeBaseHorizonSeparation)
     create_edge("edge_0", 1, "node_0", "node_1", false),
     create_edge("edge_1", 3, "node_1", "node_2", true)};
 
-  auto res = vda5050_core::order_utils::is_valid_graph(order);
+  auto res = vda5050_core::validation::is_valid_graph(order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(
@@ -275,7 +275,7 @@ TEST_F(OrderValidationTest, OrderUpdateAgainstCurrentBase)
     create_edge("edge_3", 5, "node_2", "node_4", true),
   };
 
-  auto res = vda5050_core::order_utils::is_valid_update(base_order, next_order);
+  auto res = vda5050_core::validation::is_valid_update(base_order, next_order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(res.errors.size(), 1);
@@ -303,7 +303,7 @@ TEST_F(OrderValidationTest, DisconnectedOrderUpdate)
 
   next_order.nodes = {create_node("node_4", 6, true)};
 
-  auto res = vda5050_core::order_utils::is_valid_update(base_order, next_order);
+  auto res = vda5050_core::validation::is_valid_update(base_order, next_order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(res.errors.size(), 1);
@@ -332,7 +332,7 @@ TEST_F(OrderValidationTest, StitchingNewOrder)
 
   next_order.nodes = {create_node("node_4", 4, true)};
 
-  auto res = vda5050_core::order_utils::is_valid_update(base_order, next_order);
+  auto res = vda5050_core::validation::is_valid_update(base_order, next_order);
 
   EXPECT_FALSE(res);
   EXPECT_EQ(res.errors.size(), 1);
@@ -366,7 +366,7 @@ TEST_F(OrderValidationTest, OrderValidationSuccess)
     create_edge("edge_3", 5, "node_2", "node_4", true),
   };
 
-  auto res = vda5050_core::order_utils::is_valid_update(base_order, next_order);
+  auto res = vda5050_core::validation::is_valid_update(base_order, next_order);
 
   EXPECT_TRUE(res);
 }
